Rewrite SearchBar
